Add files via upload
Before 31/12/1998 the Euro Exchange rates are theoretical ones, use the XEU code to see ECU exchange rates (
with converter in the past or historical rates page)

These theoretical rates, were computed from weight of each currency in XEU basket. As Great Britain Pound and Greek Drachma were part of the XEU basket but not member of the euro in 1999, there is a little difference between XEU and EUR (up to 2.5 % difference between 1990 and 31/12/1998). 
Before 13/03/1979 XEU rates are theoretical ones (computed from DEM, BEF/LUF, DKK, FRF, GBP, IEP, ITL and NLG rates against USD, Using weights in ECU basket when it was issued on 13/03/1979). 
If you are looking for a currency that you can't find, try converter in the past, there are more currencies available. 
Some Wikipedia pages explain how we did to compute theoretical historical rates of the euro and XEU : in English, in French, 
If you are looking for a currency that you can't find, try converter in the past, there are more currencies available.